skills config
